name and split 77480.c

+++ include/common_structs.h
@@ -27,7 +27,7 @@ typedef s32 b32;
 typedef s16 b16;
 typedef s8 b8;
 
-typedef s32 HitID;
+typedef s16 HitID;
 typedef u32 AnimID;
 
 typedef struct {
@@ -248,7 +248,7 @@ typedef struct Npc {
     /* 0x08A */ s16 yawCamOffset;
     /* 0x08C */ s16 turnAroundYawAdjustment;
     /* 0x08E */ s16 duration; // TODO: name less vaguely
-    /* 0x090 */ Vec3s homePos;
+    /* 0x090 */ Vec3f homePos;
     /* 0x096 */ s16 unk_96;
     /* 0x098 */ s16 imgfxType;
     /* 0x09A */ s16 imgfxArg1;
